+++ src/main/scala/org/podval/tools/test/detect/AnnotatedFingerprintDetector.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.test.detect
 
 import sbt.testing.{AnnotatedFingerprint, Framework}
-import xsbt.api.Discovered
 
 private final class AnnotatedFingerprintDetector(
   fingerprint: AnnotatedFingerprint,
@@ -12,5 +11,4 @@ private final class AnnotatedFingerprintDetector(
 ):
   override def name: String = fingerprint.annotationName
   override def isModule: Boolean = fingerprint.isModule
-  override protected def names(discovered: Discovered): Set[String] = discovered.annotations