+++ src/main/scala/org/podval/tools/test/detect/FingerprintDetector.scala
@@ -2,7 +2,6 @@ package org.podval.tools.test.detect
 
 import org.podval.tools.test.taskdef.Fingerprints
 import sbt.testing.{Fingerprint, Framework}
-import xsbt.api.Discovered
 
 abstract class FingerprintDetector(
   val fingerprint: Fingerprint,
@@ -17,12 +16,6 @@ abstract class FingerprintDetector(
     case _ => false
 
   final override def hashCode: Int = 37*framework.name.hashCode + fingerprint.hashCode
-  
-  final def isDiscovered(discovered: Discovered): Boolean =
-    (discovered.isModule == isModule) &&
-    names(discovered).contains(name)
-
-  protected def names(discovered: Discovered): Set[String]
       
   def isModule: Boolean