+++ src/main/scala/org/podval/tools/test/detect/SubclassFingerprintDetector.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.test.detect
 
 import sbt.testing.{Framework, SubclassFingerprint}
-import xsbt.api.Discovered
 
 final class SubclassFingerprintDetector(
   fingerprint: SubclassFingerprint,
@@ -12,5 +11,4 @@ final class SubclassFingerprintDetector(
 ):
   override def name: String = fingerprint.superclassName
   override def isModule: Boolean = fingerprint.isModule
-  override protected def names(discovered: Discovered): Set[String] = discovered.baseClasses